Database: update migration and seeder for adresses

+++ database/migrations/2022_03_20_200324_create_addresses_table.php
@@ -29,8 +29,8 @@ return new class extends Migration
             $table->text('description')->nullable();
             $table->json('details')->nullable();
             $table->uuid('subcategory_slug');
-            $table->integer('osm_id', false)->nullable();
-            $table->integer('osm_place_id', false)->nullable();
+            $table->bigInteger('osm_id', false)->nullable();
+            $table->bigInteger('osm_place_id', false)->nullable();
             $table->string('gmap_pluscode', 255)->nullable();
             $table->timestamps();
             $table->softDeletes();
